Fix: Income vs Expenses report going blank for some users (#3442)
* refactor(types): Introduce explicit NormalizedNetIncome type

- Allows for defining a single item of the array returned from
_normalizeNetIncomes if need-be
- Allows for explicitly typing the return value of _normalizeNetIncomes
to ensure that if we change the code within it, we will know it will
still return the same type

* refactor(IncomeVsExpenseComponent): Extract normalizer to a utility function

By breaking it out to a standalone utility function, we can more easily
write and run unit tests against this function without needing to deal
with rendering the whole component just to test out this function

* test(normalizeNetIncomes): Add initial tests

- Adds some todos for tests for general functionality
- Adds tests to check the edge cases where a user may have just expenses
  or just incomes in some early months, causing the length of the
  expenses and incomes arrays to differ
- Creates a MockDateWithoutTime class that partially implements
  DateWithoutTime so we can create some mock dates that will work to pass
  to the test.
- Creates a generator function to encapsulate the act of creating
  MockDateWithoutTime's and using them to create MonthyTotals to use in
  tests

* test(normalizeNetIncomes): Add more edge case tests, rework existing ones

- Adds edge case tests covering potential situations where one list may
  start earlier and end earlier than the other.
- Reworks existing edge case tests to show the work of calculating
  expected months and improve the expect statements

* fix(normalizeNetIncomes): Handle edge cases

There are some instances where expenses and incomes monthly totals don't
start or end on the same months. The case where they might start on
different months can happen if you create an account, then add some
transactions to it with dates in prior months to the month the starting
balance was set it. The case where they might end on different months may
not actually exist, since latest months should always have at least an
empty array of totals for both. But it can't hurt to handle it just incase.

* fix(MonthlySavingsRatioRow): Add fallback to 0 if totals for a given month don't exist

* docs(README): Use consistent indentation, fix numbered list item

Using markdown preview, you can see that the unordered lists that are
not indented or only indented 2 spaces register at the same level as
the order list they're contained in. 3 or 4 spaces of indentation make
them register as 1 level deeper, within the current ordered list item
they're defined in. Since other places in the file seem to use a even
number of spaces for indentation, let's make these ordered list items
have a consistent indentation level of 4 spaces.

5 was used twice in the ordered list item. So the second one is now a
6. Also removes the extra space between the number and the item text.